Add: API to allow blocks to access global styles.
