+++ packages/block-editor/src/utils/index.js
@@ -1,3 +1,7 @@
 export { default as transformStyles } from './transform-styles';
 export * from './theme';
 export * from './block-variation-transforms';
+export {
+	getResolvedStyleVariable as __experimentalGetResolvedStyleVariable,
+	getPresetVariableRepresentingAValue as __experimentalGetPresetVariableRepresentingAValue,
+} from './style-variable-resolution';
